requirements.txt
google-cloud-storage
google-cloud-secret-manager
sendgrid
flask
requests 